tools/custom_models/upload_file_to_data_registry/custom_model/requirements.txt
boto3>=1.36.24
opentelemetry-instrumentation-botocore==0.54b1
opentelemetry-instrumentation-requests==0.54b1
